Build using my Jaguar SDK Makefile helpers
Clean up the Makefile a bit and include the
standard SDK Makefile rules and definitions.

Differences in behavior:

-Will always assemble and link with -g (Add source
 level debugging data) and link with -l (add local
 symbols) now since those are the defaults in the
 SDK Makefile helper code. Edit the default value
 for LINKFLAGS in ${JAGSDK}/tools/build/jagdefs.mk
 to disable them. Note -g has no effect with rmac/
 rln.

-Won't build with with full verbosity by default.
 Use the "V" variable on the command line, e.g.:

   $ make V=2

 to restore that behavior if desired.

-Will build with rmac/rln rather than mac/aln by
 default. Set ASM=mac LINK=aln on the make command
 line or edit ${JAGSDK}/tools/build/jagdefs.mk to
 build with the Atari tools (on an OS where they
 work) intead.

Other than that, everything should be identical.
When taking the steps recommended above to tweak
the build behavior manually, the jagmand.cof and
jagbjl.cof created by this build are identical to
skunkbios_3.0.2.cof and startbjl.cof. When using
rmac/rln instead, the files are identical besides
some slight differences at the very end that
appear to be some extra metadata or symbol table
entries that aln must append, so the result is
functionally equivalent.

Tested by flashing a jagmand.cof built with this
Makefile to my skunkboard and using it a bit.