Use skunklib from Jaguar SDK
Rather than including skunk.s directly from a
relative path, build it separately and link it,
using the standard skunklib helpers from the
Jaguar SDK.

Bump the alignment up to phrase out of paranoia,
since skunk.s was previously included on a phrase
boundary. I don't know that there is any code in
skunk.s that requires this, but better safe than
sorry. This introduces slight code diffs between
the checked-in 3.0.2 bios binary and the
jagmand.cof built from this Makefile due to the
slight change in location of a few symbols, but
there should be no functional difference. There
was no affect on the jagbjl.cof build with this
change.

Tested by flashing a jagmand.cof built with these
changes to my skunkboard and using it a bit.