Default to the 93C46 EEPROM on Rev.5 boards
By default, the EZHost GPIO25 and GPIO28 pins used
to select between the 93C46 and 93C86 serial
EEPROMs on the Rev. 5 skunkboards are floating or
undefined: GPIO25 is configured as an input and
GPIO28 is in use by the UART hardware. To make any
use of the serial EEPROMs on these boards, those
GPIO pins must be configured as outputs and one
(and only one) of them must be set high. This can
be done using JCP after the board has booted and
before loading or launching any code/ROMs, but
ideally the board would boot up with at least one
of the serial EEPROMs usable, so the BIOS will
need to be able to do this as well.

However, writing to the registers that control
these GPIOs and disable the UART to free up GPIO28
can't be done directly via the usual HPI DMA mode
used to access other EZHost memory. An EZHost LPC
command must be sent over the HPI mailbox instead.
This change implements that LPC command and uses
it to enable the 93C46 chip at boot.