Preserve e2p choice across flash and error check
Before resetting the EZHost after flashing a ROM,
save off the current value of the GPIO1 output
data control register that selects the EEPROM and
restore it at the end of the reset operation. This
ensures the serial EEPROM enabled before the flash
operation (if any) is still selected after the
flash operation.

Doing this required adding a routine to read
EZHost control registers in addition to writing
them. Luckily it is almost identical to the write
routine, so almost all the code can be reused in
place.

A routine to error check both of these commands
and generate some user feedback when they fail is
also added and called any time one of the EZHost
control registers is read or written.